+++ math/isprime.cpp
@@ -11,14 +11,12 @@
 using namespace std;
 #endif
 
-#include "../pre/macros.cpp"
-
-bool isprime(ll n) {
+bool isprime(long long n) {
 	if (n<2) return false;
 	if (n==2) return true;
 	if (n%2==0) return false;
 
-	for (ll i=3; i*i<=n; i+=2) if (n%i==0) return false;
+	for (long long i=3; i*i<=n; i+=2) if (n%i==0) return false;
 
 	return true;
 }
